update userinfo get

+++ src/main/java/club/piclight/libraryman/Repository/UserInfoRepository.java
@@ -4,6 +4,6 @@ import club.piclight.libraryman.Model.UserInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
-    UserInfo getUserInfoById(Integer id);
+    UserInfo getUserInfoById(Long id);
     UserInfo getUserINfoByUserName(String name);
 }
